# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MSPR HealthAI Coach - AI Nutrition",
    "description": "\nMicro-service d'analyse nutritionnelle et de generation de plans repas par IA, partie de la plateforme MSPR HealthAI Coach.\n\nDeux pipelines d'IA sont exposes :\n\n- **Analyse de repas** : photo en entree, classification HuggingFace (`nateraw/food`, modele Food-101), lookup nutritionnel sur les datasets ETL, detection de desequilibres macros vs profil utilisateur, puis recommandations generees par Ollama (Gemma3:4b) avec fallback matrice statique.\n- **Generation de plans repas** : objectif de sante, preferences alimentaires et budget, prompt structure soumis a Ollama, validation du JSON et persistance du plan sur 1 a 30 jours.\n\nLe service expose egalement la gestion du profil nutritionnel (`nutrition-goals/me`) et l'historique pagine des analyses et des plans (`meal-analyses/me`, `meal-plans/me`). L'authentification se fait via JWT signe par MSPR-AUTH (`Authorization: Bearer <jwt>`).\n",
    "contact": {
      "name": "Arthur Poncin",
      "email": "arthur.poncin@sixense-group.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analyze-meal": {
      "post": {
        "tags": [
          "Analyse"
        ],
        "summary": "Analyse une photo de repas (macros + recommandations IA)",
        "description": "Analyse une photo de repas et renvoie macros, desequilibres et recommandations.\n\nPipeline applique :\n\n1. Classification par le modele HuggingFace `nateraw/food` (Food-101, top-3 aliments + confiance).\n2. Lookup nutritionnel sur la table `nutrition_entries` (datasets ETL).\n3. Detection des desequilibres macros (calories, proteines, glucides, lipides) par rapport au profil utilisateur (`nutrition-goals/me`). Si aucun profil n'est configure, l'analyse renvoie quand meme les macros sans recommandations.\n4. Generation des recommandations par Ollama (Gemma3:4b) avec cache 30 jours par hash (top_label, health_goal, imbalances). En cas d'echec LLM, repli sur la matrice statique 16 cas (`fallback: true`).\n5. Persistance dans `meal_analyses` pour consultation via `GET /api/v1/meal-analyses/me`.\n\n**Limitations** : modele entraine sur Food-101 (101 classes occidentales), inference CPU (1-3 s par appel). Pour les recommandations, le LLM peut etre lent (5-15 s sur CPU), prevoir un timeout cote client.\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis (JWT signe par MSPR-AUTH).",
        "operationId": "analyze_meal_api_v1_analyze_meal_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_meal_api_v1_analyze_meal_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analyse reussie : macros, desequilibres et recommandations.",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "analysis_id": 137,
                  "detected_foods": [
                    {
                      "label": "pizza",
                      "confidence": 0.85,
                      "nutrition": {
                        "calories": 1300,
                        "protein_g": 30,
                        "carbs_g": 160,
                        "fat_g": 50,
                        "fiber_g": 5
                      }
                    },
                    {
                      "label": "lasagna",
                      "confidence": 0.07
                    }
                  ],
                  "macros": {
                    "calories": 1300,
                    "protein_g": 30,
                    "carbs_g": 160,
                    "fat_g": 50,
                    "fiber_g": 5
                  },
                  "imbalances": [
                    "Apport calorique eleve (65 % du daily target).",
                    "Glucides au-dessus de la cible journaliere."
                  ],
                  "recommendations": [
                    "Reduis la portion au prochain repas et privilegie une assiette plus protee.",
                    "Equilibre la journee avec un diner riche en legumes verts."
                  ],
                  "fallback": false
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Image illisible, corrompue ou aucun aliment detecte."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          },
          "413": {
            "description": "Image trop volumineuse (limite 10 Mo)."
          },
          "415": {
            "description": "Type MIME non supporte (utiliser JPEG, PNG ou WebP)."
          }
        }
      }
    },
    "/api/v1/meal-analyses/me": {
      "get": {
        "tags": [
          "Historique"
        ],
        "summary": "Historique pagine des analyses de l'utilisateur",
        "description": "Renvoie l'historique pagine des analyses de repas pour l'utilisateur authentifie.\n\nTri decroissant par date d'analyse. Chaque element contient les aliments detectes, les macros, les recommandations sauvegardees et la date.\n\nPagination par `limit` (defaut 20, max 100) et `offset` (defaut 0). Le total est inclus dans la reponse pour permettre au client d'afficher un compteur ou de preparer une pagination infinie.\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis. L'historique est strictement scope a l'utilisateur du JWT.",
        "operationId": "list_my_meal_analyses_api_v1_meal_analyses_me_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste pagine d'analyses (tri decroissant sur created_at).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAnalysesResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": 137,
                      "detected_foods": [
                        {
                          "label": "pizza",
                          "confidence": 0.85,
                          "nutrition": {
                            "calories": 1300,
                            "protein_g": 30,
                            "carbs_g": 160,
                            "fat_g": 50
                          }
                        }
                      ],
                      "macros": {
                        "calories": 1300,
                        "protein_g": 30,
                        "carbs_g": 160,
                        "fat_g": 50,
                        "fiber_g": 5
                      },
                      "recommendations": [
                        "Reduis la portion au prochain repas."
                      ],
                      "created_at": "2026-04-29T10:15:00"
                    }
                  ],
                  "total": 137,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Validation Pydantic en echec (type ou contrainte de champ violee)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          }
        }
      }
    },
...
    "/api/v1/nutrition-goals/me": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Recupere le profil nutritionnel de l'utilisateur",
        "description": "Renvoie le profil nutritionnel de l'utilisateur authentifie.\n\nLe profil regroupe l'objectif sante (`weight_loss`, `muscle_gain`, `balance`, `sport_performance`), les cibles caloriques et macros, les allergies declarees et le regime (`omnivore`, `vegetarien`, `vegan`, `sans_gluten`).\n\nRenvoie 404 si aucun profil n'a encore ete configure (un PUT initial est alors necessaire).\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis. Le profil est strictement scope a l'utilisateur du JWT.",
        "operationId": "get_my_profile_api_v1_nutrition_goals_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profil nutritionnel courant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionGoalResponse"
                },
                "example": {
                  "user_id": 42,
                  "health_goal": "muscle_gain",
                  "calories_target": 2400,
                  "protein_g": "140.0",
                  "carbs_g": "260.0",
                  "fat_g": "80.0",
                  "allergies": [
                    "arachides",
                    "fruits a coque"
                  ],
                  "diet_type": "omnivore"
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Validation Pydantic en echec (type ou contrainte de champ violee)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          },
          "404": {
            "description": "Aucun profil nutritionnel configure (faire un PUT initial)."
          }
        }
      },
      "put": {
        "tags": [
          "Profil"
        ],
        "summary": "Cree ou met a jour le profil nutritionnel (upsert)",
        "description": "Cree ou met a jour le profil nutritionnel de l'utilisateur authentifie (upsert).\n\nTous les champs sont optionnels : un PUT partiel ecrase la ligne avec la valeur fournie. Pour conserver une valeur existante, le client doit la renvoyer explicitement.\n\nLes valeurs sont utilisees ensuite par `/analyze-meal` (detection des desequilibres) et `/generate-meal-plan` (resolution de l'objectif sante par defaut).\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis.",
        "operationId": "upsert_my_profile_api_v1_nutrition_goals_me_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionGoalRequest"
              },
              "examples": {
                "muscle_gain": {
                  "summary": "Prise de masse, omnivore",
                  "value": {
                    "health_goal": "muscle_gain",
                    "calories_target": 2400,
                    "protein_g": 140,
                    "carbs_g": 260,
                    "fat_g": 80,
                    "allergies": [
                      "arachides"
                    ],
                    "diet_type": "omnivore"
                  }
                },
                "weight_loss_vegetarien": {
                  "summary": "Perte de poids, vegetarien",
                  "value": {
                    "health_goal": "weight_loss",
                    "calories_target": 1700,
                    "protein_g": 90,
                    "carbs_g": 170,
                    "fat_g": 55,
                    "allergies": [],
                    "diet_type": "vegetarien"
                  }
                },
                "minimal": {
                  "summary": "Profil minimal (objectif sante seul)",
                  "value": {
                    "health_goal": "balance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil mis a jour (ou cree).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionGoalResponse"
                },
                "example": {
                  "user_id": 42,
                  "health_goal": "muscle_gain",
                  "calories_target": 2400,
                  "protein_g": "140.0",
                  "carbs_g": "260.0",
                  "fat_g": "80.0",
                  "allergies": [
                    "arachides",
                    "fruits a coque"
                  ],
                  "diet_type": "omnivore"
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Payload invalide (objectif sante inconnu, valeurs non numeriques, etc)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityLevel": {
        "type": "string",
        "enum": [
          "sedentary",
          "light",
          "moderate",
          "active",
          "very_active"
        ],
        "title": "ActivityLevel"
      },
      "Body_analyze_meal_api_v1_analyze_meal_post": {
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "title": "Photo",
            "description": "Photo du repas (JPEG, PNG ou WebP, 10 Mo max)."
          },
          "meal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MealType"
              },
              {
                "type": "null"
              }
            ],
            "description": "breakfast / lunch / dinner / snack. Defaut : fallback TDEE/4."
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_analyze_meal_api_v1_analyze_meal_post"
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female"
        ],
        "title": "Gender"
      },
...
      "HealthGoal": {
        "type": "string",
        "enum": [
          "weight_loss",
          "muscle_gain",
          "balance",
          "sport_performance"
        ],
        "title": "HealthGoal"
      },
...
      "MealAnalysisItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "detected_foods": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Detected Foods"
          },
          "macros": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Macros"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "detected_foods",
          "macros",
          "recommendations",
          "created_at"
        ],
        "title": "MealAnalysisItem"
      },
...
      "MealType": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack"
        ],
        "title": "MealType"
      },
      "NutritionGoalRequest": {
        "properties": {
          "health_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HealthGoal"
              },
              {
                "type": "null"
              }
            ]
          },
          "calories_target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Target"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies",
            "default": []
          },
          "diet_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diet Type"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "activity_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "NutritionGoalRequest"
      },
      "NutritionGoalResponse": {
        "properties": {
          "health_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HealthGoal"
              },
              {
                "type": "null"
              }
            ]
          },
          "calories_target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Target"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies",
            "default": []
          },
          "diet_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diet Type"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "activity_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "NutritionGoalResponse"
      },
      "PaginatedAnalysesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealAnalysisItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedAnalysesResponse",
        "example": {
          "items": [
            {
              "created_at": "2026-04-29T10:15:00",
              "detected_foods": [
                {
                  "confidence": 0.85,
                  "label": "pizza",
                  "nutrition": {}
                }
              ],
              "id": 42,
              "macros": {
                "calories": 1300,
                "carbs_g": 160,
                "fat_g": 50,
                "protein_g": 30
              },
              "recommendations": [
                "Reduis la portion au prochain repas."
              ]
            }
          ],
          "limit": 20,
          "offset": 0,
          "total": 137
        }
      },
...
    }
  },
...
}